requirements.txt
pandas==1.5.3
streamlit==1.20.0
Pillow==9.4.0
folium==0.14.0
streamlit-folium==0.11.1
inflection==0.5.1